# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AsistenRequest:
    properties:
      foto_url:
        type: string
      medsos_link:
        type: string
      nama:
        type: string
      nim:
        type: string
      nomor_hp:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - nama
    - nim
    type: object
...
  dto.BulkResponse:
    properties:
      total_inserted:
        type: integer
      total_processed:
        type: integer
      total_updated:
        type: integer
    type: object
...
  dto.UserBulkRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/dto.UserRequest'
        type: array
    required:
    - users
    type: object
  dto.UserRequest:
    properties:
      gelombang:
        enum:
        - 1
        - 2
        type: integer
      kelas_id:
        type: integer
      kelompok:
        type: string
      nama:
        type: string
      nim:
        type: string
      shift:
        enum:
        - 1
        - 2
        type: integer
    required:
    - nama
    - nim
    type: object
...
  entity.Kelas:
    properties:
      id:
        type: integer
      nama_kelas:
        type: string
    type: object
...
  entity.RoleType:
    enum:
    - user
    - admin
    - superadmin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
    - RoleSuperAdmin
...
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      foto_url:
        type: string
      gelombang:
        description: hanya dipakai saat ujian praktik
        type: integer
      id:
        type: integer
      is_registered:
        type: boolean
      kelas:
        $ref: '#/definitions/entity.Kelas'
      kelas_id:
        type: integer
      kelompok:
        type: string
      last_login_at:
        type: string
      medsos_link:
        type: string
      nama:
        type: string
      nim:
        type: string
      nomor_hp:
        type: string
      role:
        $ref: '#/definitions/entity.RoleType'
      shift:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: naufalraihans@example.com
    name: Developer
  description: REST API Documentation for Web Lab-AP v3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lab-AP API
  version: "1.0"
paths:
...
  /admin/asisten:
    get:
      description: Mengambil daftar seluruh asisten
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Daftar Asisten
      tags:
      - Info
      - Admin - Asisten
    post:
      consumes:
      - application/json
      description: Menambahkan akun asisten baru
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AsistenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
      security:
      - bearerAuth: []
      summary: Tambah Asisten
      tags:
      - Admin - Asisten
  /admin/asisten/{id}:
    put:
      consumes:
      - application/json
      description: Mengubah data akun asisten
      parameters:
      - description: ID Asisten
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AsistenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
      security:
      - bearerAuth: []
      summary: Perbarui Asisten
      tags:
      - Admin - Asisten
...
  /admin/users:
    get:
      description: Mengambil data mahasiswa dengan filter opsional (kelas_id, shift)
      parameters:
      - description: ID Kelas
        in: query
        name: kelas_id
        type: integer
      - description: Shift
        in: query
        name: shift
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Daftar Mahasiswa
      tags:
      - Admin - User
    post:
      consumes:
      - application/json
      description: Menambahkan data mahasiswa baru
      parameters:
      - description: Payload Mahasiswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
      security:
      - bearerAuth: []
      summary: Tambah Mahasiswa
      tags:
      - Admin - User
  /admin/users/{id}:
    delete:
      description: Menghapus pengguna (mahasiswa) dari sistem
      parameters:
      - description: ID User
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - bearerAuth: []
      summary: Hapus User
      tags:
      - Admin - User
    put:
      consumes:
      - application/json
      description: Memperbarui data mahasiswa (termasuk reset perangkat jika diatur)
      parameters:
      - description: ID Mahasiswa
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Mahasiswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
      security:
      - bearerAuth: []
      summary: Perbarui Mahasiswa
      tags:
      - Admin - User
  /admin/users/{id}/reset-password:
    post:
      description: Me-reset password mahasiswa (kosong). Mahasiswa harus register
        ulang.
      parameters:
      - description: ID Mahasiswa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - bearerAuth: []
      summary: Reset Password
      tags:
      - Admin - User
  /admin/users/bulk:
    post:
      consumes:
      - application/json
      description: Menambahkan atau memperbarui banyak data mahasiswa sekaligus dari
        JSON (biasanya hasil parsing CSV di Frontend). Jika NIM sudah ada, data akan
        di-update.
      parameters:
      - description: Payload Bulk Mahasiswa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserBulkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/dto.BulkResponse'
              type: object
      security:
      - bearerAuth: []
      summary: Import Mahasiswa (Bulk)
      tags:
      - Admin - User
...
  /info/asisten:
    get:
      description: Mengambil daftar seluruh asisten
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Daftar Asisten
      tags:
      - Info
      - Admin - Asisten
...
swagger: "2.0"